Trouble connecting through sockets
More error points, socket sending strings updated to deal with info

+++ StrategyGame/Tile.cpp
@@ -54,7 +54,7 @@ Building* Tile::building() {
     return this->tileBuilding;
 }
 
-float Tile::terrain() {
+unsigned int Tile::terrain() {
     return this->tileTerrain;
 }
